Adds activity posting for other social activities.
We now place a "follow" etc posts in our activities aggregate. I also
added a function to determine the human description for the activity.
It's a bit confusing to grok it, but the function attempts to give some
standard components to put together the language. Localization efforts
should be focused there.